requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-debug-toolbar==3.2
psycopg2==2.8.6
pytz==2021.1
sentry-sdk==1.0.0
sqlparse==0.4.1
urllib3==1.26.4
